+++ src/lib.rs
@@ -116,7 +116,7 @@ mod storage;
 pub use self::storage::BlockType;
 
 mod traits;
-pub use self::traits::{Bits, BitsExt, BitsMut, BitsPush, BitSliceable};
+pub use self::traits::{BitRange, Bits, BitsExt, BitsMut, BitsPush, BitSliceable};
 
 mod slice;
 pub use self::slice::{BitSlice, BitSliceMut};